Rename command to clone:create, Add clone:sharing, clone:config, clone:features

+++ spec/spec_helper.rb
@@ -8,25 +8,13 @@ module HerokuHelpers
 
   MOCK = ENV['MOCK'] != 'false'
 
-  def data_site_crt
-    @data_site_crt ||= File.read(File.join(DATA_PATH, 'site.crt'))
-  end
-
-  def data_site_key
-    @data_site_key ||= File.read(File.join(DATA_PATH, 'site.key'))
-  end
-
   def heroku
     # ENV['HEROKU_API_KEY'] used for :api_key
     Heroku::API.new(:mock => MOCK)
   end
 
-  def random_domain
-    "#{random_name}.com"
-  end
-
   def random_name
-    "akira-#{SecureRandom.hex(10)}"
+    "myapp-#{SecureRandom.hex(10)}"
   end
 
   def random_email_address
@@ -43,35 +31,43 @@ module HerokuHelpers
     with_blank_git_repository do
       begin
         data = heroku.post_app(random_params.merge(params)).body
-        @name = data['name']
+        name = data['name']
 
-        heroku.post_collaborator(@name, random_email_address)
-        heroku.put_config_vars(@name, random_name.upcase => random_name)
+        heroku.post_collaborator(name, random_email_address)
+        heroku.put_config_vars(name, random_name.upcase => random_name)
 
         ready = false
         until ready
-          ready = heroku.request(:method => :put, :path => "/apps/#{@name}/status").status == 201
+          ready = heroku.request(:method => :put, :path => "/apps/#{name}/status").status == 201
         end
         yield(data)
       ensure
-        heroku.delete_app(@name) rescue nil
+        heroku.delete_app(name) rescue nil
       end
     end
   end
 
   def with_blank_git_repository(&block)
-    sandbox = File.join(Dir.tmpdir, "heroku", Process.pid.to_s)
-    FileUtils.mkdir_p(sandbox)
+    if @has_blank_git_repo
+      block.call
+    else
+      begin
+        sandbox = File.join(Dir.tmpdir, "heroku", Process.pid.to_s)
+        FileUtils.mkdir_p(sandbox)
 
-    old_dir = Dir.pwd
-    Dir.chdir(sandbox)
+        old_dir = Dir.pwd
+        Dir.chdir(sandbox)
 
-    `git init`
-    block.call
+        `git init`
+        @has_blank_git_repo = true
+        block.call
 
-    FileUtils.rm_rf(sandbox)
-  ensure
-    Dir.chdir(old_dir)
+        FileUtils.rm_rf(sandbox)
+        @has_blank_git_repo = false
+      ensure
+        Dir.chdir(old_dir)
+      end
+    end
   end
 
   def execute(command_line)
